Add support for very large portfolios (#333)
This code change does several things, breaking public APIs in the process:

- Changes the data type of participation/investment/transaction entity IDs from int to long, as it's now only a matter of time until Zonky outgrows int.
- Changes all data types from int to long where the totals have to do with portfolio size, which has no upper limit. (This is mostly theoretical, but the institutional investor may already be hitting Integer.MAX_VALUE.)

+++ robozonky-api/src/main/java/com/github/robozonky/api/remote/ControlApi.java
@@ -54,11 +54,11 @@ public interface ControlApi {
 
     @POST
     @Path("/smp/investments/{id}/shares")
-    void purchase(@PathParam("id") int id, PurchaseRequest purchaseRequest);
+    void purchase(@PathParam("id") long id, PurchaseRequest purchaseRequest);
 
     @DELETE
     @Path("/traded-investments/{id}")
-    void cancel(@PathParam("id") int id);
+    void cancel(@PathParam("id") long id);
 
 }
