+++ robozonky-common/src/test/java/com/github/robozonky/util/NumberUtilTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 The RoboZonky Project
+ * Copyright 2018 The RoboZonky Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,8 +20,8 @@ import java.math.BigDecimal;
 
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.assertj.core.api.SoftAssertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 class NumberUtilTest {
 
@@ -35,8 +35,8 @@ class NumberUtilTest {
 
     @Test
     void hasAdditions() {
-        final int[] original = new int[]{1};
-        final int[] updated = new int[]{1, 2};
+        final long[] original = new long[]{1};
+        final long[] updated = new long[]{1, 2};
         assertSoftly(softly -> {
             softly.assertThat(NumberUtil.hasAdditions(original, updated)).isTrue();
             softly.assertThat(NumberUtil.hasAdditions(original, original)).isFalse();
