+++ robozonky-notifications/src/main/java/com/github/robozonky/notifications/listeners/InvestmentPurchasedEventListener.java
@@ -44,7 +44,7 @@ public class InvestmentPurchasedEventListener extends AbstractListener<Investmen
     protected Map<String, Object> getData(final InvestmentPurchasedEvent event) {
         final Investment i = event.getInvestment();
         final Map<String, Object> result = super.getData(event);
-        final int invested = event.getPortfolioOverview().getCzkInvested().intValue();
+        final long invested = event.getPortfolioOverview().getCzkInvested().longValue();
         result.put("yield", FinancialCalculator.expectedInterestAfterFees(i, invested));
         final BigDecimal interestRate = FinancialCalculator.expectedInterestRateAfterFees(i, invested);
         result.put("relativeYield", interestRate);
